+++ tests/Dynamic/GainCompressorFilter.cpp
@@ -17,69 +17,69 @@ constexpr gsl::index PROCESSSIZE = 64;
 
 BOOST_AUTO_TEST_CASE( GainFilter_threshold_test )
 {
-  ATK::GainFilter<ATK::GainCompressorFilter<float>> filter;
+  ATK::GainFilter<ATK::GainCompressorFilter<double>> filter;
   filter.set_threshold(10);
   BOOST_CHECK_EQUAL(filter.get_threshold(), 10);
 }
 
 BOOST_AUTO_TEST_CASE( GainFilter_threshold_db_test )
 {
-  ATK::GainFilter<ATK::GainCompressorFilter<float>> filter;
+  ATK::GainFilter<ATK::GainCompressorFilter<double>> filter;
   filter.set_threshold_db(20);
   BOOST_CHECK_EQUAL(filter.get_threshold(), 100);
 }
 
 BOOST_AUTO_TEST_CASE( GainFilter_threshold_range_test )
 {
-  ATK::GainFilter<ATK::GainCompressorFilter<float>> filter;
+  ATK::GainFilter<ATK::GainCompressorFilter<double>> filter;
   BOOST_CHECK_THROW(filter.set_threshold(0), std::out_of_range);
 }
 
 BOOST_AUTO_TEST_CASE( GainFilter_ratio_test )
 {
-  ATK::GainFilter<ATK::GainCompressorFilter<float>> filter;
+  ATK::GainFilter<ATK::GainCompressorFilter<double>> filter;
   filter.set_ratio(10);
   BOOST_CHECK_EQUAL(filter.get_ratio(), 10);
 }
 
 BOOST_AUTO_TEST_CASE( GainFilter_ratio_range_test )
 {
-  ATK::GainFilter<ATK::GainCompressorFilter<float>> filter;
+  ATK::GainFilter<ATK::GainCompressorFilter<double>> filter;
   BOOST_CHECK_THROW(filter.set_ratio(0), std::out_of_range);
 }
 
 BOOST_AUTO_TEST_CASE( GainCompressorFilter_softness_test )
 {
-  ATK::GainFilter<ATK::GainCompressorFilter<float>> filter;
+  ATK::GainFilter<ATK::GainCompressorFilter<double>> filter;
   filter.set_softness(0.5);
   BOOST_CHECK_EQUAL(filter.get_softness(), 0.5);
 }
 
 BOOST_AUTO_TEST_CASE( GainCompressorFilter_softness_range_test )
 {
-  ATK::GainFilter<ATK::GainCompressorFilter<float>> filter;
+  ATK::GainFilter<ATK::GainCompressorFilter<double>> filter;
   BOOST_CHECK_THROW(filter.set_softness(-0.000001), std::out_of_range);
 }
 
 BOOST_AUTO_TEST_CASE( GainCompressorFilter_const_1_test )
 {
-  std::array<float, PROCESSSIZE> data;
+  std::array<double, PROCESSSIZE> data;
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
     data[i] = 1;
   }
   
-  ATK::InPointerFilter<float> generator(data.data(), 1, PROCESSSIZE, false);
+  ATK::InPointerFilter<double> generator(data.data(), 1, PROCESSSIZE, false);
   generator.set_output_sampling_rate(48000);
 
-  std::array<float, PROCESSSIZE> outdata;
+  std::array<double, PROCESSSIZE> outdata;
 
-  ATK::GainFilter<ATK::GainCompressorFilter<float>> filter;
+  ATK::GainFilter<ATK::GainCompressorFilter<double>> filter;
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
   filter.set_threshold(10);
 
-  ATK::OutPointerFilter<float> output(outdata.data(), 1, PROCESSSIZE, false);
+  ATK::OutPointerFilter<double> output(outdata.data(), 1, PROCESSSIZE, false);
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &filter, 0);
 
@@ -93,22 +93,22 @@ BOOST_AUTO_TEST_CASE( GainCompressorFilter_const_1_test )
 
 BOOST_AUTO_TEST_CASE( GainCompressorFilter_const_0_test )
 {
-  std::array<float, PROCESSSIZE> data;
+  std::array<double, PROCESSSIZE> data;
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
     data[i] = 0;
   }
 
-  ATK::InPointerFilter<float> generator(data.data(), 1, PROCESSSIZE, false);
+  ATK::InPointerFilter<double> generator(data.data(), 1, PROCESSSIZE, false);
   generator.set_output_sampling_rate(48000);
 
-  std::array<float, PROCESSSIZE> outdata;
+  std::array<double, PROCESSSIZE> outdata;
 
-  ATK::GainFilter<ATK::GainCompressorFilter<float>> filter(1);
+  ATK::GainFilter<ATK::GainCompressorFilter<double>> filter(1);
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
 
-  ATK::OutPointerFilter<float> output(outdata.data(), 1, PROCESSSIZE, false);
+  ATK::OutPointerFilter<double> output(outdata.data(), 1, PROCESSSIZE, false);
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &filter, 0);
 
@@ -122,25 +122,25 @@ BOOST_AUTO_TEST_CASE( GainCompressorFilter_const_0_test )
 
 BOOST_AUTO_TEST_CASE( GainCompressorFilter_const_1_threshold_05_ratio_2_test )
 {
-  std::array<float, PROCESSSIZE> data;
+  std::array<double, PROCESSSIZE> data;
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
     data[i] = 1;
   }
 
-  ATK::InPointerFilter<float> generator(data.data(), 1, PROCESSSIZE, false);
+  ATK::InPointerFilter<double> generator(data.data(), 1, PROCESSSIZE, false);
   generator.set_output_sampling_rate(48000);
 
-  std::array<float, PROCESSSIZE> outdata;
+  std::array<double, PROCESSSIZE> outdata;
 
-  ATK::GainFilter<ATK::GainCompressorFilter<float>> filter(1);
+  ATK::GainFilter<ATK::GainCompressorFilter<double>> filter(1);
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
   filter.set_threshold(0.5);
   filter.set_ratio(2);
   filter.set_softness(1);
 
-  ATK::OutPointerFilter<float> output(outdata.data(), 1, PROCESSSIZE, false);
+  ATK::OutPointerFilter<double> output(outdata.data(), 1, PROCESSSIZE, false);
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &filter, 0);
 
@@ -154,25 +154,25 @@ BOOST_AUTO_TEST_CASE( GainCompressorFilter_const_1_threshold_05_ratio_2_test )
 
 BOOST_AUTO_TEST_CASE( GainCompressorFilter_const_1_threshold_05_ratio_4_test )
 {
-  std::array<float, PROCESSSIZE> data;
+  std::array<double, PROCESSSIZE> data;
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
     data[i] = 1;
   }
 
-  ATK::InPointerFilter<float> generator(data.data(), 1, PROCESSSIZE, false);
+  ATK::InPointerFilter<double> generator(data.data(), 1, PROCESSSIZE, false);
   generator.set_output_sampling_rate(48000);
 
-  std::array<float, PROCESSSIZE> outdata;
+  std::array<double, PROCESSSIZE> outdata;
 
-  ATK::GainFilter<ATK::GainCompressorFilter<float>> filter(1);
+  ATK::GainFilter<ATK::GainCompressorFilter<double>> filter(1);
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
   filter.set_threshold(0.5);
   filter.set_ratio(4);
   filter.set_softness(1);
 
-  ATK::OutPointerFilter<float> output(outdata.data(), 1, PROCESSSIZE, false);
+  ATK::OutPointerFilter<double> output(outdata.data(), 1, PROCESSSIZE, false);
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &filter, 0);
