+++ tests/Dynamic/GainExpanderFilter.cpp
@@ -17,35 +17,35 @@ constexpr gsl::index PROCESSSIZE = 64;
 
 BOOST_AUTO_TEST_CASE( GainExpanderFilter_softness_test )
 {
-  ATK::GainFilter<ATK::GainExpanderFilter<float>> filter;
+  ATK::GainFilter<ATK::GainExpanderFilter<double>> filter;
   filter.set_softness(0.5);
   BOOST_CHECK_EQUAL(filter.get_softness(), 0.5);
 }
 
 BOOST_AUTO_TEST_CASE( GainExpanderFilter_softness_range_test )
 {
-  ATK::GainFilter<ATK::GainExpanderFilter<float>> filter;
+  ATK::GainFilter<ATK::GainExpanderFilter<double>> filter;
   BOOST_CHECK_THROW(filter.set_softness(-0.000001), std::out_of_range);
 }
 
 BOOST_AUTO_TEST_CASE( GainExpanderFilter_const_1_test )
 {
-  std::array<float, PROCESSSIZE> data;
+  std::array<double, PROCESSSIZE> data;
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
     data[i] = 1;
   }
   
-  ATK::InPointerFilter<float> generator(data.data(), 1, PROCESSSIZE, false);
+  ATK::InPointerFilter<double> generator(data.data(), 1, PROCESSSIZE, false);
   generator.set_output_sampling_rate(48000);
 
-  std::array<float, PROCESSSIZE> outdata;
+  std::array<double, PROCESSSIZE> outdata;
 
-  ATK::GainFilter<ATK::GainExpanderFilter<float>> filter(1);
+  ATK::GainFilter<ATK::GainExpanderFilter<double>> filter(1);
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
 
-  ATK::OutPointerFilter<float> output(outdata.data(), 1, PROCESSSIZE, false);
+  ATK::OutPointerFilter<double> output(outdata.data(), 1, PROCESSSIZE, false);
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &filter, 0);
 
@@ -59,24 +59,24 @@ BOOST_AUTO_TEST_CASE( GainExpanderFilter_const_1_test )
 
 BOOST_AUTO_TEST_CASE( GainExpanderFilter_const_0_test )
 {
-  std::array<float, PROCESSSIZE> data;
+  std::array<double, PROCESSSIZE> data;
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
     data[i] = 0;
   }
 
-  ATK::InPointerFilter<float> generator(data.data(), 1, PROCESSSIZE, false);
+  ATK::InPointerFilter<double> generator(data.data(), 1, PROCESSSIZE, false);
   generator.set_output_sampling_rate(48000);
 
-  std::array<float, PROCESSSIZE> outdata;
+  std::array<double, PROCESSSIZE> outdata;
 
-  ATK::GainFilter<ATK::GainExpanderFilter<float>> filter(1);
+  ATK::GainFilter<ATK::GainExpanderFilter<double>> filter(1);
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
   filter.set_threshold(0.1);
   filter.set_ratio(2);
 
-  ATK::OutPointerFilter<float> output(outdata.data(), 1, PROCESSSIZE, false);
+  ATK::OutPointerFilter<double> output(outdata.data(), 1, PROCESSSIZE, false);
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &filter, 0);
 
@@ -84,31 +84,31 @@ BOOST_AUTO_TEST_CASE( GainExpanderFilter_const_0_test )
 
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
-    BOOST_REQUIRE_SMALL(outdata[i], 0.001f);
+    BOOST_REQUIRE_SMALL(outdata[i], 0.001);
   }
 }
 
 BOOST_AUTO_TEST_CASE( GainExpanderFilter_const_1_threshold_2_ratio_2_test )
 {
-  std::array<float, PROCESSSIZE> data;
+  std::array<double, PROCESSSIZE> data;
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
     data[i] = 1;
   }
 
-  ATK::InPointerFilter<float> generator(data.data(), 1, PROCESSSIZE, false);
+  ATK::InPointerFilter<double> generator(data.data(), 1, PROCESSSIZE, false);
   generator.set_output_sampling_rate(48000);
 
-  std::array<float, PROCESSSIZE> outdata;
+  std::array<double, PROCESSSIZE> outdata;
 
-  ATK::GainFilter<ATK::GainExpanderFilter<float>> filter(1);
+  ATK::GainFilter<ATK::GainExpanderFilter<double>> filter(1);
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
   filter.set_threshold(2);
   filter.set_ratio(2);
   filter.set_softness(1);
 
-  ATK::OutPointerFilter<float> output(outdata.data(), 1, PROCESSSIZE, false);
+  ATK::OutPointerFilter<double> output(outdata.data(), 1, PROCESSSIZE, false);
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &filter, 0);
 
@@ -122,25 +122,25 @@ BOOST_AUTO_TEST_CASE( GainExpanderFilter_const_1_threshold_2_ratio_2_test )
 
 BOOST_AUTO_TEST_CASE( GainExpanderFilter_const_1_threshold_2_ratio_4_test )
 {
-  std::array<float, PROCESSSIZE> data;
+  std::array<double, PROCESSSIZE> data;
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
     data[i] = 1;
   }
 
-  ATK::InPointerFilter<float> generator(data.data(), 1, PROCESSSIZE, false);
+  ATK::InPointerFilter<double> generator(data.data(), 1, PROCESSSIZE, false);
   generator.set_output_sampling_rate(48000);
 
-  std::array<float, PROCESSSIZE> outdata;
+  std::array<double, PROCESSSIZE> outdata;
 
-  ATK::GainFilter<ATK::GainExpanderFilter<float>> filter(1);
+  ATK::GainFilter<ATK::GainExpanderFilter<double>> filter(1);
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
   filter.set_threshold(2);
   filter.set_ratio(4);
   filter.set_softness(1);
 
-  ATK::OutPointerFilter<float> output(outdata.data(), 1, PROCESSSIZE, false);
+  ATK::OutPointerFilter<double> output(outdata.data(), 1, PROCESSSIZE, false);
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &filter, 0);
 
@@ -154,25 +154,25 @@ BOOST_AUTO_TEST_CASE( GainExpanderFilter_const_1_threshold_2_ratio_4_test )
 
 BOOST_AUTO_TEST_CASE( GainExpanderFilter_const_1_threshold_2_ratio_4_test_twice )
 {
-  std::array<float, PROCESSSIZE> data;
+  std::array<double, PROCESSSIZE> data;
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
     data[i] = 1;
   }
   
-  ATK::InPointerFilter<float> generator(data.data(), 1, PROCESSSIZE, false);
+  ATK::InPointerFilter<double> generator(data.data(), 1, PROCESSSIZE, false);
   generator.set_output_sampling_rate(48000);
   
-  std::array<float, PROCESSSIZE> outdata;
+  std::array<double, PROCESSSIZE> outdata;
   
-  ATK::GainFilter<ATK::GainExpanderFilter<float>> filter(1);
+  ATK::GainFilter<ATK::GainExpanderFilter<double>> filter(1);
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
   filter.set_threshold(2);
   filter.set_ratio(4);
   filter.set_softness(1);
   
-  ATK::OutPointerFilter<float> output(outdata.data(), 1, PROCESSSIZE, false);
+  ATK::OutPointerFilter<double> output(outdata.data(), 1, PROCESSSIZE, false);
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &filter, 0);
